+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.2.0] 2023-10-16
+### Changed
+- Rename rename internal save methods from `save` and `save_all` to `_save` and `_save_all`.
+- Rename `skip` argument to `offset` on methods `list`, `filter` and `find`.
+
+
+## [0.1.0] 2023-10-15
+### Changed
+- Rename package name from `sa_model_crud` to `sa_modelcrud`.
+
+
 ## [0.0.1] 2023-10-12
 ### Added
 - Initial version.